Impl unmute
afaik there is no way to tell if a video is muted or not. I should grep
through omxplayer for dbus commands but it didn't look like it. You can
easily fetch the volume but it doesn't reflect whether it's muted.

Instead there are separate 'm' and 'u' commands to mute/unmute. This
might be better anyway because there is nonzero latency, and it's
tempting to spam commands